backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.104.0
anthropic==0.34.2
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
starlette==0.27.0
